djangoRestfulApis/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
config==0.4.2
cycler==0.10.0
Deprecated==1.2.7
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
docopt==0.6.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
psycopg2==2.8.4
PyGithub==1.46
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pythonanywhere==0.8.3
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.0
virtualenvwrapper-win==1.2.5
whitenoise==5.0.1
wrapt==1.12.0
zipp==2.2.0
